Fix FnAsArray for reverse args. Move pipe error.
